OPS-4810: Get Parsoid to stop properly on ctrl+c or SIGINT by porting changes from upstream
While working on this I discovered that original upstream changes (very fresh) were not correct - Parsoid still wasn't stopping properly on ctrl+c - this is because ctrl+c sends SIGINT not only to parent process but also to all foreground processes. I worked with Arlolra from WMF and now this is fixed.